move bindings_generator/ to tool/generator/
