+++ tool/generator/banned_names.dart
