+++ tool/generator/generate_bindings.dart
