+++ tool/generator/package.json
