+++ tool/generator/singletons.dart
