+++ tool/generator/translator.dart
